Create ErrorCollector class; put Station into separate module; eliminate Station->City dependency
