+++ subway_structure.py
@@ -1,16 +1,43 @@
 from __future__ import annotations
 
-import math
 import re
 from collections import Counter, defaultdict
 from collections.abc import Callable, Collection, Iterator
 from itertools import chain, islice
-from typing import TypeAlias, TypeVar
+from typing import TypeVar
 
 from css_colours import normalize_colour
+from structure.consts import (
+    CONSTRUCTION_KEYS,
+    DEFAULT_MODES_OVERGROUND,
+    DEFAULT_MODES_RAPID,
+    MAX_DISTANCE_STOP_TO_LINE,
+    RAILWAY_TYPES,
+)
+from structure.error_collector import (
+    ErrorCollectorGate,
+    ErrorCollectorHolder,
+)
+from structure.geom_utils import (
+    angle_between,
+    distance,
+    find_segment,
+    distance_on_line,
+    project_on_line,
+)
+from structure.osm_element import el_center, el_id
+from structure.station import Station
+from structure.types import (
+    CriticalValidationError,
+    IdT,
+    LonLat,
+    OsmElementT,
+    RailT,
+    TransfersT,
+    TransferT,
+)
 
 MAX_DISTANCE_TO_ENTRANCES = 300  # in meters
-MAX_DISTANCE_STOP_TO_LINE = 50  # in meters
 ALLOWED_STATIONS_MISMATCH = 0.02  # part of total station count
 ALLOWED_TRANSFERS_MISMATCH = 0.07  # part of total interchanges count
 ALLOWED_ANGLE_BETWEEN_STOPS = 45  # in degrees
@@ -21,38 +48,12 @@ SUGGEST_TRANSFER_MIN_DISTANCE = 100  # in meters
 # it is likely the same object
 DISPLACEMENT_TOLERANCE = 300  # in meters
 
-MODES_RAPID = {"subway", "light_rail", "monorail", "train"}
-MODES_OVERGROUND = {"tram", "bus", "trolleybus", "aerialway", "ferry"}
-DEFAULT_MODES_RAPID = {"subway", "light_rail"}
-DEFAULT_MODES_OVERGROUND = {"tram"}  # TODO: bus and trolleybus?
-ALL_MODES = MODES_RAPID | MODES_OVERGROUND
-RAILWAY_TYPES = {
-    "rail",
-    "light_rail",
-    "subway",
-    "narrow_gauge",
-    "funicular",
-    "monorail",
-    "tram",
-}
-CONSTRUCTION_KEYS = (
-    "construction",
-    "proposed",
-    "construction:railway",
-    "proposed:railway",
-)
 
 used_entrances = set()
 
 
 START_END_TIMES_RE = re.compile(r".*?(\d{2}):(\d{2})-(\d{2}):(\d{2}).*")
 
-OsmElementT: TypeAlias = dict
-IdT: TypeAlias = str  # Type of feature ids
-TransferT: TypeAlias = set[IdT]  # A transfer is a set of StopArea IDs
-TransfersT: TypeAlias = list[TransferT]
-LonLat: TypeAlias = tuple[float, float]
-RailT: TypeAlias = list[LonLat]
 T = TypeVar("T")
 
 
@@ -97,262 +98,11 @@ def osm_interval_to_seconds(interval_str: str) -> int | None:
     return seconds + 60 * minutes + 60 * 60 * hours
 
 
-class CriticalValidationError(Exception):
-    """Is thrown if an error occurs
-    that prevents further validation of a city."""
-
-
-def el_id(el: OsmElementT) -> IdT | None:
-    if not el:
-        return None
-    if "type" not in el:
-        raise Exception("What is this element? {}".format(el))
-    return el["type"][0] + str(el.get("id", el.get("ref", "")))
-
-
-def el_center(el: OsmElementT) -> LonLat | None:
-    if not el:
-        return None
-    if "lat" in el:
-        return el["lon"], el["lat"]
-    elif "center" in el:
-        return el["center"]["lon"], el["center"]["lat"]
-    return None
-
-
-def distance(p1: LonLat, p2: LonLat) -> float:
-    if p1 is None or p2 is None:
-        raise Exception(
-            "One of arguments to distance({}, {}) is None".format(p1, p2)
-        )
-    dx = math.radians(p1[0] - p2[0]) * math.cos(
-        0.5 * math.radians(p1[1] + p2[1])
-    )
-    dy = math.radians(p1[1] - p2[1])
-    return 6378137 * math.sqrt(dx * dx + dy * dy)
-
-
-def is_near(p1: LonLat, p2: LonLat) -> bool:
-    return (
-        p1[0] - 1e-8 <= p2[0] <= p1[0] + 1e-8
-        and p1[1] - 1e-8 <= p2[1] <= p1[1] + 1e-8
-    )
-
-
-def project_on_segment(p: LonLat, p1: LonLat, p2: LonLat) -> float | None:
-    """Given three points, return u - the position of projection of
-    point p onto segment p1p2 regarding point p1 and (p2-p1) direction vector
-    """
-    dp = (p2[0] - p1[0], p2[1] - p1[1])
-    d2 = dp[0] * dp[0] + dp[1] * dp[1]
-    if d2 < 1e-14:
-        return None
-    u = ((p[0] - p1[0]) * dp[0] + (p[1] - p1[1]) * dp[1]) / d2
-    if not 0 <= u <= 1:
-        return None
-    return u
-
-
-def project_on_line(p: LonLat, line: RailT) -> dict:
-    result = {
-        # In the first approximation, position on rails is the index of the
-        # closest vertex of line to the point p. Fractional value means that
-        # the projected point lies on a segment between two vertices.
-        # More than one value can occur if a route follows the same tracks
-        # more than once.
-        "positions_on_line": None,
-        "projected_point": None,  # (lon, lat)
-    }
-
-    if len(line) < 2:
-        return result
-    d_min = MAX_DISTANCE_STOP_TO_LINE * 5
-    closest_to_vertex = False
-    # First, check vertices in the line
-    for i, vertex in enumerate(line):
-        d = distance(p, vertex)
-        if d < d_min:
-            result["positions_on_line"] = [i]
-            result["projected_point"] = vertex
-            d_min = d
-            closest_to_vertex = True
-        elif vertex == result["projected_point"]:
-            # Repeated occurrence of the track vertex in line, like Oslo Line 5
-            result["positions_on_line"].append(i)
-    # And then calculate distances to each segment
-    for seg in range(len(line) - 1):
-        # Check bbox for speed
-        if not (
-            (
-                min(line[seg][0], line[seg + 1][0]) - MAX_DISTANCE_STOP_TO_LINE
-                <= p[0]
-                <= max(line[seg][0], line[seg + 1][0])
-                + MAX_DISTANCE_STOP_TO_LINE
-            )
-            and (
-                min(line[seg][1], line[seg + 1][1]) - MAX_DISTANCE_STOP_TO_LINE
-                <= p[1]
-                <= max(line[seg][1], line[seg + 1][1])
-                + MAX_DISTANCE_STOP_TO_LINE
-            )
-        ):
-            continue
-        u = project_on_segment(p, line[seg], line[seg + 1])
-        if u:
-            projected_point = (
-                line[seg][0] + u * (line[seg + 1][0] - line[seg][0]),
-                line[seg][1] + u * (line[seg + 1][1] - line[seg][1]),
-            )
-            d = distance(p, projected_point)
-            if d < d_min:
-                result["positions_on_line"] = [seg + u]
-                result["projected_point"] = projected_point
-                d_min = d
-                closest_to_vertex = False
-            elif projected_point == result["projected_point"]:
-                # Repeated occurrence of the track segment in line,
-                # like Oslo Line 5
-                if not closest_to_vertex:
-                    result["positions_on_line"].append(seg + u)
-    return result
-
-
-def find_segment(
-    p: LonLat, line: RailT, start_vertex: int = 0
-) -> tuple[int, float] | tuple[None, None]:
-    """Returns index of a segment and a position inside it."""
-    EPS = 1e-9
-    for seg in range(start_vertex, len(line) - 1):
-        if is_near(p, line[seg]):
-            return seg, 0.0
-        if line[seg][0] == line[seg + 1][0]:
-            if not (p[0] - EPS <= line[seg][0] <= p[0] + EPS):
-                continue
-            px = None
-        else:
-            px = (p[0] - line[seg][0]) / (line[seg + 1][0] - line[seg][0])
-        if px is None or (0 <= px <= 1):
-            if line[seg][1] == line[seg + 1][1]:
-                if not (p[1] - EPS <= line[seg][1] <= p[1] + EPS):
-                    continue
-                py = None
-            else:
-                py = (p[1] - line[seg][1]) / (line[seg + 1][1] - line[seg][1])
-            if py is None or (0 <= py <= 1):
-                if py is None or px is None or (px - EPS <= py <= px + EPS):
-                    return seg, px or py
-    return None, None
-
-
-def distance_on_line(
-    p1: LonLat, p2: LonLat, line: RailT, start_vertex: int = 0
-) -> tuple[float, int] | None:
-    """Calculates distance via line between projections
-    of points p1 and p2. Returns a TUPLE of (d, vertex):
-    d is the distance and vertex is the number of the second
-    vertex, to continue calculations for the next point."""
-    line_len = len(line)
-    seg1, pos1 = find_segment(p1, line, start_vertex)
-    if seg1 is None:
-        # logging.warn('p1 %s is not projected, st=%s', p1, start_vertex)
-        return None
-    seg2, pos2 = find_segment(p2, line, seg1)
-    if seg2 is None:
-        if line[0] == line[-1]:
-            line = line + line[1:]
-            seg2, pos2 = find_segment(p2, line, seg1)
-        if seg2 is None:
-            # logging.warn('p2 %s is not projected, st=%s', p2, start_vertex)
-            return None
-    if seg1 == seg2:
-        return distance(line[seg1], line[seg1 + 1]) * abs(pos2 - pos1), seg1
-    if seg2 < seg1:
-        # Should not happen
-        raise Exception("Pos1 %s is after pos2 %s", seg1, seg2)
-    d = 0
-    if pos1 < 1:
-        d += distance(line[seg1], line[seg1 + 1]) * (1 - pos1)
-    for i in range(seg1 + 1, seg2):
-        d += distance(line[i], line[i + 1])
-    if pos2 > 0:
-        d += distance(line[seg2], line[seg2 + 1]) * pos2
-    return d, seg2 % line_len
-
-
-def angle_between(p1: LonLat, c: LonLat, p2: LonLat) -> float:
-    a = round(
-        abs(
-            math.degrees(
-                math.atan2(p1[1] - c[1], p1[0] - c[0])
-                - math.atan2(p2[1] - c[1], p2[0] - c[0])
-            )
-        )
-    )
-    return a if a <= 180 else 360 - a
-
-
 def format_elid_list(ids: Collection[IdT]) -> str:
     msg = ", ".join(sorted(ids)[:20])
     if len(ids) > 20:
         msg += ", ..."
     return msg
-
-
-class Station:
-    @staticmethod
-    def get_modes(el: OsmElementT) -> set[str]:
-        modes = {m for m in ALL_MODES if el["tags"].get(m) == "yes"}
-        if mode := el["tags"].get("station"):
-            modes.add(mode)
-        return modes
-
-    @staticmethod
-    def is_station(el: OsmElementT, modes: set[str]) -> bool:
-        # public_transport=station is too ambiguous and unspecific to use,
-        # so we expect for it to be backed by railway=station.
-        if (
-            "tram" in modes
-            and el.get("tags", {}).get("railway") == "tram_stop"
-        ):
-            return True
-        if el.get("tags", {}).get("railway") not in ("station", "halt"):
-            return False
-        for k in CONSTRUCTION_KEYS:
-            if k in el["tags"]:
-                return False
-        # Not checking for station=train, obviously
-        if "train" not in modes and Station.get_modes(el).isdisjoint(modes):
-            return False
-        return True
-
-    def __init__(self, el: OsmElementT, city: City) -> None:
-        """Call this with a railway=station node."""
-        if not Station.is_station(el, city.modes):
-            raise Exception(
-                "Station object should be instantiated from a station node. "
-                "Got: {}".format(el)
-            )
-
-        self.id: IdT = el_id(el)
-        self.element: OsmElementT = el
-        self.modes = Station.get_modes(el)
-        self.name = el["tags"].get("name", "?")
-        self.int_name = el["tags"].get(
-            "int_name", el["tags"].get("name:en", None)
-        )
-        try:
-            self.colour = normalize_colour(el["tags"].get("colour", None))
-        except ValueError as e:
-            self.colour = None
-            city.warn(str(e), el)
-        self.center = el_center(el)
-        if self.center is None:
-            raise Exception("Could not find center of {}".format(el))
-
-    def __repr__(self) -> str:
-        return "Station(id={}, modes={}, name={}, center={})".format(
-            self.id, ",".join(self.modes), self.name, self.center
-        )
 
 
 class StopArea:
@@ -1912,14 +1662,12 @@ class RouteMaster:
         )
 
 
-class City:
+class City(ErrorCollectorHolder, ErrorCollectorGate):
     route_class = Route
 
     def __init__(self, city_data: dict, overground: bool = False) -> None:
+        ErrorCollectorHolder.__init__(self)
         self.validate_called = False
-        self.errors: list[str] = []
-        self.warnings: list[str] = []
-        self.notices: list[str] = []
         self.id = None
         self.try_fill_int_attribute(city_data, "id")
         self.name = city_data["name"]
@@ -2011,23 +1759,6 @@ class City:
                 tags.get("name", tags.get("ref", "")),
             )
         return message
-
-    def notice(self, message: str, el: OsmElementT | None = None) -> None:
-        """This type of message may point to a potential problem."""
-        msg = City.log_message(message, el)
-        self.notices.append(msg)
-
-    def warn(self, message: str, el: OsmElementT | None = None) -> None:
-        """A warning is definitely a problem but is doesn't prevent
-        from building a routing file and doesn't invalidate the city.
-        """
-        msg = City.log_message(message, el)
-        self.warnings.append(msg)
-
-    def error(self, message: str, el: OsmElementT | None = None) -> None:
-        """Error is a critical problem that invalidates the city."""
-        msg = City.log_message(message, el)
-        self.errors.append(msg)
 
     def contains(self, el: OsmElementT) -> bool:
         center = el_center(el)
@@ -2230,12 +1961,12 @@ class City:
 
     @property
     def is_good(self) -> bool:
-        if not (self.errors or self.validate_called):
+        if not (self.has_errors or self.validate_called):
             raise RuntimeError(
                 "You mustn't refer to City.is_good property before calling "
                 "the City.validate() method unless an error already occurred."
             )
-        return len(self.errors) == 0
+        return not self.has_errors
 
     def get_validation_result(self) -> dict:
         result = {
